added path for wsl in windows

+++ shell/zsh/.zim/zimfw.zsh
@@ -4,7 +4,7 @@
 # MIT License
 #
 # Copyright (c) 2015-2016 Matt Hamilton and contributors
-# Copyright (c) 2016-2021 Eric Nielsen, Matt Hamilton and contributors
+# Copyright (c) 2016-2022 Eric Nielsen, Matt Hamilton and contributors
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -46,8 +46,7 @@ _zimfw_mv() {
     if [[ -e ${2} ]]; then
       command mv -f ${2}{,.old} || return 1
     fi
-    command mv -f ${1} ${2} && \
-        _zimfw_print -PR "%F{green})%f %B${2}:%b Updated. Restart your terminal for changes to take effect."
+    command mv -f ${1} ${2} && _zimfw_print -PR "%F{green})%f %B${2}:%b Updated.${_zrestartmsg}"
   fi
 }
 
@@ -62,45 +61,21 @@ _zimfw_build_init() {
     print -R "zmodule() { source ${ZIM_HOME}/zimfw.zsh \"\${@}\" }"
     # Remove all prefixes from _zfpaths, _zfunctions and _zcmds
     local -r zpre=$'*\0'
-    if (( ${#_zfpaths} )) print -R 'fpath=('${${_zfpaths#${~zpre}}:A}' ${fpath})'
+    print -R 'typeset -g _zim_fpath=('${${_zfpaths#${~zpre}}:A}')'
+    if (( ${#_zfpaths} )) print 'fpath=(${_zim_fpath} ${fpath})'
     if (( ${#_zfunctions} )) print -R 'autoload -Uz -- '${_zfunctions#${~zpre}}
     print -R ${(F)_zcmds#${~zpre}}
   ) ${ztarget}
 }
 
 _zimfw_build_login_init() {
-  # Array with unique dirs. ${ZIM_HOME} or any subdirectory should only occur once.
-  local -Ur zscriptdirs=(${ZIM_HOME} ${${_zdirs##${ZIM_HOME}/*}:A})
-  local -r zscriptglob=("${^zscriptdirs[@]}/(^*test*/)#*.zsh(|-theme)(N-.)") ztarget=${ZIM_HOME}/login_init.zsh
+  local -r ztarget=${ZIM_HOME}/login_init.zsh
   # Force update of login_init.zsh if it's older than .zimrc
   if [[ ${ztarget} -ot ${ZDOTDIR:-${HOME}}/.zimrc ]]; then
     command mv -f ${ztarget}{,.old} || return 1
   fi
   _zimfw_mv =(
-    print -nR "() {
-  builtin emulate -L zsh
-  setopt EXTENDED_GLOB
-  autoload -Uz zrecompile
-  local zdumpfile zfile
-
-  # Compile the completion cache; significant speedup
-  zstyle -s ':zim:completion' dumpfile 'zdumpfile' || zdumpfile=\${ZDOTDIR:-\${HOME}}/.zcompdump
-  if [[ -f \${zdumpfile} ]]; then
-    zrecompile -p \${1} \${zdumpfile} || return 1
-  fi
-
-  # Compile Zsh startup files
-  for zfile in \${ZDOTDIR:-\${HOME}}/.z(shenv|profile|shrc|login|logout)(N-.); do
-    zrecompile -p \${1} \${zfile} || return 1
-  done
-
-  # Compile Zim scripts
-  for zfile in ${zscriptglob}; do
-    zrecompile -p \${1} \${zfile} || return 1
-  done
-
-  if [[ \${1} != -q ]] print -P 'Done with compile.'
-} \"\${@}\"
+    print -nR "# Do nothing. This file is deprecated.
 "
   ) ${ztarget}
 }
@@ -133,6 +108,7 @@ Repository options:
                              %Bdegit%b requires curl or wget, and currently only works with GitHub
                              URLs. Modules install faster and take less disk space. Local
                              changes are lost on updates. Git submodules are not supported.
+  %B--no-submodules%b            Don't install or update git submodules.
   %B-z%b|%B--frozen%b                Don't install or update the module.
 
 Initialization options:
@@ -142,7 +118,7 @@ Initialization options:
                              %Bfunctions%b subdirectory, if any.
   %B-s%b|%B--source%b <file_path>    Source specified file. The file path is relative to the module
                              root directory. Default: %Binit.zsh%b, if the %Bfunctions%b subdirectory
-                             also exists, or the file with largest size and with name matching
+                             also exists, or the largest of the files with name matching
                              %B{init.zsh,module_name.{zsh,plugin.zsh,zsh-theme,sh}}%b, if any.
   %B-c%b|%B--cmd%b <command>         Execute specified command. Occurrences of the %B{}%b placeholder in
                              the command are substituted by the module root directory path.
@@ -162,7 +138,7 @@ Initialization options:
     return 2
   fi
   local zurl=${1} zmodule=${1:t} ztool zdir ztype zrev zarg
-  local -i zdisabled=0 zfrozen=0
+  local -i zsubmodules=1 zfrozen=0 zdisabled=0
   local -a zfpaths zfunctions zcmds
   zstyle -s ':zim:zmodule' use 'ztool' || ztool=git
   if [[ ${zurl} =~ ^[^:/]+: ]]; then
@@ -202,7 +178,7 @@ Initialization options:
         ;;
     esac
     case ${1} in
-      -b|--branch|-t|--tag|-u|--use)
+      -b|--branch|-t|--tag|-u|--use|--no-submodules)
         if [[ -z ${zurl} ]] _zimfw_print -u2 -PR "%F{yellow}! ${funcfiletrace[1]}:%B${zmodule}:%b The zmodule option %B${1}%b has no effect for external modules%f"
         ;;
     esac
@@ -221,6 +197,7 @@ Initialization options:
         shift
         ztool=${1}
         ;;
+      --no-submodules) zsubmodules=0 ;;
       -z|--frozen) zfrozen=1 ;;
       -f|--fpath)
         shift
@@ -252,7 +229,7 @@ Initialization options:
     shift
   done
   if (( _zflags & 1 )); then
-    _zmodules_zargs+=("${ztool}" "${_zargs_action}" "${zmodule}" "${zdir}" "${zurl}" "${ztype}" "${zrev}" "${zfrozen}" "${zdisabled}")
+    _zmodules_zargs+=("${ztool}" "${_zargs_action}" "${zmodule}" "${zdir}" "${zurl}" "${ztype}" "${zrev}" "${zsubmodules}" "${zfrozen}" "${zdisabled}")
   fi
   if (( _zflags & 2 )); then
     if (( zdisabled )); then
@@ -275,11 +252,11 @@ Initialization options:
         local -ra prezto_scripts=(${zdir}/init.zsh(N))
         if (( ${#zfpaths} && ${#prezto_scripts} )); then
           # this follows the prezto module format, no need to check for other scripts
-          zcmds=("source ${^prezto_scripts[@]:A}")
+          zcmds=('source '${^prezto_scripts:A})
         else
           # get script with largest size (descending `O`rder by `L`ength, and return only `[1]` first)
           local -ra zscripts=(${zdir}/(init.zsh|${zmodule:t}.(zsh|plugin.zsh|zsh-theme|sh))(NOL[1]))
-          zcmds=("source ${^zscripts[@]:A}")
+          zcmds=('source '${^zscripts:A})
         fi
       fi
       if (( ! ${#zfpaths} && ! ${#zfunctions} && ! ${#zcmds} )); then
@@ -347,41 +324,69 @@ _zimfw_version_check() {
   fi
 }
 
+_zimfw_check_dumpfile() {
+  local zdumpfile zfpath zline
+  zstyle -s ':zim:completion' dumpfile 'zdumpfile' || zdumpfile=${ZDOTDIR:-${HOME}}/.zcompdump
+  if [[ -e ${zdumpfile} ]]; then
+    if (( ${+_zim_fpath} )); then
+      local -r zpre=$'*\0'
+      zfpath=(${${_zfpaths#${~zpre}}:A} ${fpath:|_zim_fpath})
+    else
+      zfpath=(${fpath})
+    fi
+    local -r zcomps=(${^zfpath}/^([^_]*|*~|*.zwc(|.old))(N:t))
+    IFS=$' \t' read -rA zline < ${zdumpfile} || return 1
+    if [[ ${zline[2]} -eq ${#zcomps} && ${zline[4]} == ${ZSH_VERSION} ]]; then
+      _zimfw_print -PR "%F{green})%f %B${zdumpfile}:%b Already up to date"
+    else
+      _zimfw_print -PR "%F{green})%f %B${zdumpfile}:%b New completion configuration needs to be dumped. Will do %Bclean-dumpfile%b."
+      _zimfw_clean_dumpfile
+    fi
+  else
+    _zimfw_print -PR "%F{green})%f %B${zdumpfile}:%b Not found"
+  fi
+  _zimfw_print 'Done with check-dumpfile.'
+}
+
 _zimfw_clean_compiled() {
   # Array with unique dirs. ${ZIM_HOME} or any subdirectory should only occur once.
   local -Ur zscriptdirs=(${ZIM_HOME} ${${_zdirs##${ZIM_HOME}/*}:A})
   local zopt
   if (( _zprintlevel > 0 )) zopt=-v
-  command rm -f ${zopt} ${^zscriptdirs}/**/*.zwc(|.old)(N) || return 1
-  command rm -f ${zopt} ${ZDOTDIR:-${HOME}}/.z(shenv|profile|shrc|login|logout).zwc(|.old)(N) || return 1
-  _zimfw_print -P 'Done with clean-compiled. Restart your terminal or run %Bzimfw compile%b to re-compile.'
+  command rm -f ${zopt} ${^zscriptdirs}/**/*.zwc(|.old)(N) && \
+      _zimfw_print -P 'Done with clean-compiled. Restart your terminal or run %Bzimfw compile%b to re-compile.'
 }
 
 _zimfw_clean_dumpfile() {
   local zdumpfile zopt
   zstyle -s ':zim:completion' dumpfile 'zdumpfile' || zdumpfile=${ZDOTDIR:-${HOME}}/.zcompdump
   if (( _zprintlevel > 0 )) zopt=-v
-  command rm -f ${zopt} ${zdumpfile}(|.zwc(|.old))(N) || return 1
-  _zimfw_print -P 'Done with clean-dumpfile. Restart your terminal to dump an updated configuration.'
+  command rm -f ${zopt} ${zdumpfile}(|.zwc(|.old))(N) && \
+      _zimfw_print -P "Done with clean-dumpfile.${_zrestartmsg}"
 }
 
 _zimfw_compile() {
-  local zopt
-  if (( _zprintlevel <= 0 )) zopt=-q
-  source ${ZIM_HOME}/login_init.zsh ${zopt}
+  # Array with unique dirs. ${ZIM_HOME} or any subdirectory should only occur once.
+  local -Ur zscriptdirs=(${ZIM_HOME} ${${_zdirs##${ZIM_HOME}/*}:A})
+  # Compile Zim scripts
+  local zfile
+  for zfile in ${^zscriptdirs}/(^*test*/)#*.zsh(|-theme)(N-.); do
+    if [[ ! ${zfile}.zwc -nt ${zfile} ]]; then
+      zcompile -R ${zfile} && _zimfw_print -PR "%F{green})%f %B${zfile}.zwc:%b Compiled"
+    fi
+  done
+  _zimfw_print -P 'Done with compile.'
 }
 
 _zimfw_info() {
-  print -R 'zimfw version: '${_zversion}' (built at 2021-11-21 19:40:05 UTC, previous commit is a5fb148)'
+  print -R 'zimfw version: '${_zversion}' (built at 2022-01-25 22:16:46 UTC, previous commit is 72fe20c)'
   print -R 'ZIM_HOME:      '${ZIM_HOME}
   print -R 'Zsh version:   '${ZSH_VERSION}
   print -R 'System info:   '$(command uname -a)
 }
 
 _zimfw_install_update() {
-  _zimfw_source_zimrc 1 ${1} && \
-      zargs -n 9 -P 0 -- "${_zmodules_zargs[@]}" -- _zimfw_run_tool && \
-      _zimfw_print -PR "Done with ${1}. Restart your terminal for any changes to take effect."
+  _zimfw_source_zimrc 1 ${1} && zargs -n 10 -P 0 -- "${_zmodules_zargs[@]}" -- _zimfw_run_tool
 }
 
 _zimfw_uninstall() {
@@ -423,7 +428,7 @@ _zimfw_upgrade() {
 
 _zimfw_run_list() {
   local -r ztool=${1} zmodule=${3} zdir=${4} zurl=${5} ztype=${6} zrev=${7}
-  local -ri zfrozen=${8} zdisabled=${9}
+  local -ri zsubmodules=${8} zfrozen=${9} zdisabled=${10}
   print -PnR "%B${zmodule}:%b ${zdir}"
   if [[ -z ${zurl} ]] print -Pn ' (external)'
   if (( ${zfrozen} )) print -Pn ' (frozen)'
@@ -437,14 +442,16 @@ _zimfw_run_list() {
       else
         print -nR "${ztype} ${zrev}"
       fi
-      print -R ", using ${ztool}"
+      print -nR ", using ${ztool}"
+      if (( ! zsubmodules )) print -nR ', no git submodules'
+      print
     fi
     # Match and remove the current module prefix from _zfpaths, _zfunctions and _zcmds
     local -r zpre=${zmodule}$'\0'
     local -r zfpaths=(${${(M)_zfpaths:#${zpre}*}#${zpre}}) zfunctions=(${${(M)_zfunctions:#${zpre}*}#${zpre}}) zcmds=(${${(M)_zcmds:#${zpre}*}#${zpre}})
-    if (( ${#zfpaths} )) print -R "  fpath: ${zfpaths[@]}"
-    if (( ${#zfunctions} )) print -R "  autoload: ${zfunctions[@]}"
-    if (( ${#zcmds} )) print -R "  cmd: ${(j:; :)zcmds[@]}"
+    if (( ${#zfpaths} )) print -R '  fpath: '${zfpaths}
+    if (( ${#zfunctions} )) print -R '  autoload: '${zfunctions}
+    if (( ${#zcmds} )) print -R '  cmd: '${(j:; :)zcmds}
   fi
 }
 
@@ -454,7 +461,7 @@ _zimfw_run_tool() {
     if (( _zprintlevel > 1 )) print -u2 -PR $'\E[2K\r'"%F{green})%f %B${zmodule}:%b Skipping external module"
     return 0
   fi
-  local -ri zfrozen=${8}
+  local -ri zfrozen=${9}
   if (( zfrozen )); then
     if (( _zprintlevel > 1 )) print -u2 -PR $'\E[2K\r'"%F{green})%f %B${zmodule}:%b Skipping frozen module"
     return 0
@@ -482,9 +489,8 @@ _zimfw_run_tool() {
   local zcmd
   case ${ztool} in
     degit) zcmd="# This runs in a new shell
-builtin emulate -L zsh
-setopt EXTENDED_GLOB
-readonly -i PRINTLEVEL=\${1}
+builtin emulate -L zsh -o EXTENDED_GLOB
+readonly -i PRINTLEVEL=\${1} SUBMODULES=\${8}
 readonly ACTION=\${2} MODULE=\${3} DIR=\${4} URL=\${5} REV=\${7} TEMP=.zdegit_\${RANDOM}
 readonly TARBALL_TARGET=\${DIR}/\${TEMP}_tarball.tar.gz INFO_TARGET=\${DIR}/.zdegit
 
@@ -495,8 +501,8 @@ print_error() {
 print_okay() {
   if (( PRINTLEVEL > 0 )); then
     local -r log=\${2:+\${(F):-  \${(f)^2}}}
-    if [[ -e \${DIR}/.gitmodules ]]; then
-      print -u2 -PlR $'\E[2K\r'\"%F{yellow}! %B\${MODULE}:%b \${(C)1}. Module contains git submodules, which are not supported by Zim's degit and were not \${1}.%f\" \${log}
+    if [[ \${SUBMODULES} -ne 0 && -e \${DIR}/.gitmodules ]]; then
+      print -u2 -PlR $'\E[2K\r'\"%F{yellow}! %B\${MODULE}:%b \${(C)1}. Module contains git submodules, which are not supported by Zim's degit and were not \${1}. Use zmodule option %B--no-submodules%b to disable this warning.%f\" \${log}
     else
       print -PlR $'\E[2K\r'\"%F{green})%f %B\${MODULE}:%b \${(C)1}\" \${log}
     fi
@@ -587,9 +593,7 @@ case \${ACTION} in
       # return 1 does not change \${TRY_BLOCK_ERROR}, only changes \${?}
       (( TRY_BLOCK_ERROR = ? ))
       command rm -f \${TARBALL_TARGET} 2>/dev/null
-      if (( TRY_BLOCK_ERROR )); then
-        command rm -rf \${DIR} 2>/dev/null
-      fi
+      if (( TRY_BLOCK_ERROR )) command rm -rf \${DIR} 2>/dev/null
     }
     ;;
   update)
@@ -626,8 +630,8 @@ esac
 " ;;
     git) zcmd="# This runs in a new shell
 builtin emulate -L zsh
-readonly -i PRINTLEVEL=\${1}
-readonly ACTION=\${2} MODULE=\${3} DIR=\${4} URL=\${5} TYPE=\${6:=branch} SUBMODULES=1
+readonly -i PRINTLEVEL=\${1} SUBMODULES=\${8}
+readonly ACTION=\${2} MODULE=\${3} DIR=\${4} URL=\${5} TYPE=\${6:=branch}
 REV=\${7}
 
 print_error() {
@@ -640,7 +644,7 @@ print_okay() {
 
 case \${ACTION} in
   install)
-    if ERR=\$(command git clone \${REV:+-b} \${REV} -q --config core.autocrlf=false \${SUBMODULES:+--recursive} -- \${URL} \${DIR} 2>&1); then
+    if ERR=\$(command git clone \${REV:+-b} \${REV} -q --config core.autocrlf=false \${\${SUBMODULES:#0}:+--recursive} -- \${URL} \${DIR} 2>&1); then
       print_okay Installed
     else
       print_error 'Error during git clone' \${ERR}
@@ -700,7 +704,7 @@ case \${ACTION} in
     else
       OUT=\"Updating to \${TYPE} \${REV}\"
     fi
-    if [[ -n \${SUBMODULES} ]]; then
+    if (( SUBMODULES )); then
       if ! ERR=\$(command git -C \${DIR} submodule update --init --recursive -q -- 2>&1); then
         print_error 'Error during git submodule update' \${ERR}
         return 1
@@ -715,30 +719,31 @@ esac
       return 1
       ;;
   esac
-  zsh -c ${zcmd} ${ztool} ${_zprintlevel} "${@[2,7]}"
+  zsh -c ${zcmd} ${ztool} ${_zprintlevel} "${@[2,8]}"
 }
 
 zimfw() {
-  builtin emulate -L zsh
-  setopt EXTENDED_GLOB
-  local -r _zversion='1.6.2' zusage="Usage: %B${0}%b <action> [%B-q%b|%B-v%b]
+  builtin emulate -L zsh -o EXTENDED_GLOB
+  local -r _zversion='1.8.0' zusage="Usage: %B${0}%b <action> [%B-q%b|%B-v%b]
 
 Actions:
   %Bbuild%b           Build %B${ZIM_HOME}/init.zsh%b and %B${ZIM_HOME}/login_init.zsh%b.
-                  Also does %Bcompile%b. Use %B-v%b to also see its output.
+                  Also does %Bcheck-dumpfile%b and %Bcompile%b. Use %B-v%b to also see their output.
+  %Bcheck-dumpfile%b  Does %Bclean-dumpfile%b if new completion configuration needs to be dumped.
   %Bclean%b           Clean all. Does both %Bclean-compiled%b and %Bclean-dumpfile%b.
   %Bclean-compiled%b  Clean Zsh compiled files.
-  %Bclean-dumpfile%b  Clean completion dump file.
+  %Bclean-dumpfile%b  Clean completion dumpfile.
   %Bcompile%b         Compile Zsh files.
   %Bhelp%b            Print this help.
   %Binfo%b            Print Zim and system info.
   %Blist%b            List all modules currently defined in %B${ZDOTDIR:-${HOME}}/.zimrc%b.
                   Use %B-v%b to also see the modules details.
-  %Binstall%b         Install new modules. Also does %Bbuild%b and %Bcompile%b. Use %B-v%b to also see their
-                  output, and see skipped modules.
+  %Binit%b            Same as %Binstall%b, but with output tailored to be used at terminal startup.
+  %Binstall%b         Install new modules. Also does %Bbuild%b, %Bcheck-dumpfile%b and %Bcompile%b. Use %B-v%b to
+                  also see their output, and see skipped modules.
   %Buninstall%b       Delete unused modules. Prompts for confirmation. Use %B-q%b for quiet uninstall.
-  %Bupdate%b          Update current modules. Also does %Bbuild%b and %Bcompile%b. Use %B-v%b to also see
-                  their output, and see skipped modules.
+  %Bupdate%b          Update current modules. Also does %Bbuild%b, %Bcheck-dumpfile%b and %Bcompile%b. Use %B-v%b
+                  to also see their output, and see skipped modules.
   %Bupgrade%b         Upgrade zimfw. Also does %Bcompile%b. Use %B-v%b to also see its output.
   %Bversion%b         Print zimfw version.
 
@@ -765,27 +770,36 @@ Options:
     _zimfw_version_check
   fi
 
+  local _zrestartmsg=' Restart your terminal for changes to take effect.'
   case ${1} in
     build)
       _zimfw_source_zimrc 2 && _zimfw_build || return 1
       (( _zprintlevel-- ))
-      _zimfw_compile
+      _zimfw_check_dumpfile && _zimfw_compile
       ;;
-    init) _zimfw_install_update install && _zimfw_source_zimrc 2 && _zimfw_build ;;
+    check-dumpfile) _zimfw_source_zimrc 2 && _zimfw_check_dumpfile ;;
     clean) _zimfw_source_zimrc 2 && _zimfw_clean_compiled && _zimfw_clean_dumpfile ;;
     clean-compiled) _zimfw_source_zimrc 2 && _zimfw_clean_compiled ;;
     clean-dumpfile) _zimfw_clean_dumpfile ;;
-    compile) _zimfw_source_zimrc 2 && _zimfw_build_login_init && _zimfw_compile ;;
+    compile) _zimfw_compile ;;
     help) print -PR ${zusage} ;;
     info) _zimfw_info ;;
     list)
-      _zimfw_source_zimrc 3 && zargs -n 9 -- "${_zmodules_zargs[@]}" -- _zimfw_run_list && \
+      _zimfw_source_zimrc 3 && zargs -n 10 -- "${_zmodules_zargs[@]}" -- _zimfw_run_list && \
           _zimfw_list_unuseds ' (unused)'
+      ;;
+    init)
+      _zrestartmsg=
+      _zimfw_install_update install || return 1
+      (( _zprintlevel-- ))
+      _zimfw_print -PR "Done with install.${_zrestartmsg}" # Only printed in verbose mode
+      _zimfw_source_zimrc 2 && _zimfw_build && _zimfw_check_dumpfile && _zimfw_compile
       ;;
     install|update)
       _zimfw_install_update ${1} || return 1
+      _zimfw_print -PR "Done with ${1}.${_zrestartmsg}"
       (( _zprintlevel-- ))
-      _zimfw_source_zimrc 2 && _zimfw_build && _zimfw_compile
+      _zimfw_source_zimrc 2 && _zimfw_build && _zimfw_check_dumpfile && _zimfw_compile
       ;;
     uninstall) _zimfw_source_zimrc 2 && _zimfw_list_unuseds && _zimfw_uninstall ;;
     upgrade)
